test(user): UserControllerTest에서 로그인 관련 테스트를 LoginControllerTest로 분리한다.
